requirements.txt
openai
python-dotenv
langchain
notebook
